[PATCH] fs: read_write: fix declaration-after-statement in read()
Kernel 4.9 is built as gnu89 (C90), which forbids declaring variables
after statements within a block. The read() path had a declaration:

    loff_t pos = file_pos_read(f.file);

following statements inside an if-block which also contains a KSU hook
guarded by #ifdef, triggering -Wdeclaration-after-statement with Clang
when -Werror is set.

Move the variable declaration to the top of the block and perform the
assignment afterwards.

This fixes:
  fs/read_write.c:600:10: error: mixing declarations and code is a C99 extension [-Werror,-Wdeclaration-after-statement]

No functional change.

Signed-off-by: Onelots <onelots@onelots.fr>